release 0.1.99 source code

+++ services/codeartspipeline/v2/model/model_show_step_outputs_response.go
@@ -10,7 +10,7 @@ import (
 type ShowStepOutputsResponse struct {
 	StepOutputs *[]OutputRespStepOutputs `json:"step_outputs,omitempty"`
 
-	CurrentSystemTime *int32 `json:"current_system_time,omitempty"`
+	CurrentSystemTime *int64 `json:"current_system_time,omitempty"`
 	HttpStatusCode    int    `json:"-"`
 }
